requirements.txt
delegator.py
tornado
